+++ server/middleware/authMiddleware.js
