+++ server/routes/myths.js
