+++ server/routes/welcome.js
