+++ server/schemas/myths.js
